Use LoadLibraryExW instead of LoadLibraryW
They are both available since Windows XP. But the only the LoadLibraryExW call
is available in Universal Windows Platform builds.

In this call we need LOAD_LIBRARY_SEARCH_USER_DIRS to make sure the value set
with SetDllDirectory() is used. It is included in
LOAD_LIBRARY_SEARCH_DEFAULT_DIRS [1] which will also need in case the DLL needs
to load some other DLLs in other locations.

[1] https://learn.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-loadlibraryexw